Moved to some member vars to Camera.hpp interface to hopefully inspire future programmers to use them.
